Expand catalog to 88 vehicles/65 parts, add fitment qualifiers, add Playwright QA harness
Sources real MagnaFlow product data (via research agents, not scraping) to grow the
demo catalog from 10 vehicles/4 parts to 88/65/154 fitment rows, ahead of a real
ACES/PIES data drop. Adds a qualifiers/fitment_qualifiers schema so fitment gated on
more than year/make/model/engine (e.g. Ram 1500 vs Ram 1500 Classic rear suspension)
prompts a clarifying question instead of guessing the wrong SKU. Adds a separate qa/
package that drives MagnaFlow's live site with Playwright to verify catalog SKUs and
spot-check real recommendations, decoupled from the request path.

+++ backend/src/db/import-aces.js
@@ -77,7 +77,7 @@ async function importACES(filePath) {
         const existing = await client.query(
           `SELECT id FROM vehicles WHERE year=$1 AND make=$2 AND model=$3
            AND COALESCE(submodel,'')=COALESCE($4,'')
-           AND COALESCE(engine_liters,0)=COALESCE($5,0)`,
+           AND COALESCE(engine_liters,0)=COALESCE($5::numeric,0)`,
           [year, make, model, submodel, engineLiters]
         );
         vehicleId = existing.rows[0]?.id;
